final touches based on greptile reviews
