+++ .gitignore
@@ -69,5 +69,5 @@ scratch.py
 
 .env
 
-# Ignoring congfiguration files generated by chainlit
+# Ignoring configuration files generated by chainlit
 .chainlit/*
